Cleaned up bind.h a bit, got rough design of subtype classes implemented, added simple array type as a test. Work I can't fathom at the moment needs to be done on compiling function calls and getting the return value
