+++ src/bind.h
@@ -1,6 +1,7 @@
 #pragma once
 #include <types.h>
 #include <builtin.h>
+#include <template_utils.hpp>
 
 #include <vector>
 #include <typeindex>
@@ -17,9 +18,10 @@
 #define fimm(f) asmjit::imm(reinterpret_cast<void*>(f))
 namespace gjs {
 	class vm_context;
-	enum class vm_register;
 	class vm_function;
 	class vm_type;
+	class type_manager;
+	enum class vm_register;
 
 	class bind_exception : public std::exception {
 		public:
@@ -32,58 +34,6 @@ namespace gjs {
 	};
 
 	namespace bind {
-		struct wrapped_class;
-	};
-
-
-	template<class T> struct remove_all { typedef T type; };
-	template<class T> struct remove_all<T*> : remove_all<T> {};
-	template<class T> struct remove_all<T&> : remove_all<T> {};
-	template<class T> struct remove_all<T&&> : remove_all<T> {};
-	template<class T> struct remove_all<T const> : remove_all<T> {};
-	template<class T> struct remove_all<T volatile> : remove_all<T> {};
-	template<class T> struct remove_all<T const volatile> : remove_all<T> {};
-	template<class T> struct remove_all<T[]> : remove_all<T> {};
-
-	template <typename T>
-	inline const char* base_type_name() {
-		return typeid(remove_all<T>::type).name();
-	}
-
-	class type_manager {
-		public:
-			type_manager(vm_context* ctx);
-			~type_manager();
-
-			vm_type* get(const std::string& internal_name);
-
-			vm_type* get(u32 id);
-
-			template <typename T>
-			vm_type* get() {
-				return get(base_type_name<T>());
-			}
-
-			vm_type* add(const std::string& name, const std::string& internal_name);
-
-			vm_type* finalize_class(bind::wrapped_class* wrapped);
-
-			std::vector<vm_type*> all();
-
-		protected:
-			friend class vm_function;
-			vm_context* m_ctx;
-			robin_hood::unordered_map<std::string, vm_type*> m_types;
-			robin_hood::unordered_map<u32, vm_type*> m_types_by_id;
-	};
-
-	namespace bind {
-		template <typename T>
-		void to_reg(vm_context* context, const T& in, u64* reg_ptr);
-
-		template <typename T>
-		void from_reg(vm_context* context, T* out, u64* reg_ptr);
-
 		struct wrapped_function {
 			wrapped_function(std::type_index ret, std::vector<std::type_index> args, const std::string& _name, const std::string& _sig)
 			: return_type(ret), arg_types(args), name(_name), sig(_sig), is_static_method(false), address(0), ret_is_ptr(false) { }
@@ -101,6 +51,42 @@ namespace gjs {
 			u64 address;
 		};
 
+		struct wrapped_class {
+			struct property {
+				property(wrapped_function* g, wrapped_function* s, std::type_index t, u64 o, u8 f) :
+					getter(g), setter(s), type(t), offset(o), flags(f) { }
+
+				wrapped_function* getter;
+				wrapped_function* setter;
+				std::type_index type;
+				u64 offset;
+				u8 flags;
+			};
+
+			wrapped_class(asmjit::JitRuntime& _rt, const std::string& _name, const std::string& _internal_name, size_t _size) :
+				rt(_rt), name(_name), internal_name(_internal_name), size(_size), ctor(nullptr), dtor(nullptr), requires_subtype(false)
+			{
+			}
+
+			~wrapped_class();
+
+			asmjit::JitRuntime& rt;
+			std::string name;
+			std::string internal_name;
+			bool requires_subtype;
+			wrapped_function* ctor;
+			robin_hood::unordered_map<std::string, wrapped_function*> methods;
+			robin_hood::unordered_map<std::string, property*> properties;
+			wrapped_function* dtor;
+			size_t size;
+		};
+
+
+		/*
+		 * Magic for wrapping functions and methods with ASM
+		 * so that they can be called with runtime generated
+		 * argument lists (x86 version, thanks to asmjit)
+		 */
 		namespace x86  {
 			using compiler = asmjit::x86::Compiler;
 			using reg = asmjit::x86::Gp;
@@ -121,6 +107,12 @@ namespace gjs {
 				(*self.*method)(args...);
 			}
 
+			template <typename T>
+			asmjit::BaseReg reg_for_type(compiler& c) {
+				if constexpr (std::is_floating_point_v<T>) return c.newXmm();
+				else return c.newGp(asmjit::Type::IdOfT<T>::kTypeId);
+			}
+
 			template <typename Ret, typename... Args>
 			struct global_function : wrapped_function {
 				typedef Ret (*func_type)(Args...);
@@ -136,49 +128,88 @@ namespace gjs {
 					if (!tpm->get<Ret>()) {
 						throw bind_exception(format("Return type '%s' of function '%s' has not been bound yet", base_type_name<Ret>(), name.c_str()));
 					}
+					// describe the function for the wrapped_function interface
 					ret_is_ptr = std::is_reference_v<Ret> || std::is_pointer_v<Ret>;
 					arg_is_ptr = { (std::is_reference_v<Args> || std::is_pointer_v<Args>)... };
 					address = u64(reinterpret_cast<void*>(f));
 					
+					// begin generating the wrapper function
 					asmjit::CodeHolder h;
 					h.init(rt.codeInfo());
 					compiler c(&h);
+
+					// define function signature 'void wrapped(void* out_result, void** args)'
 					c.addFunc(asmjit::FuncSignatureT<void, void*, void**>(asmjit::CallConv::kIdHost));
+
+					// bind out_result to register
 					reg out = c.newGp(asmjit::Type::IdOfT<Ret*>::kTypeId);
-					reg in_args = c.newGp(asmjit::Type::IdOfT<void**>::kTypeId);
 					c.setArg(0, out);
+
+					// bind args to register
+					reg in_args = c.newGp(asmjit::Type::IdOfT<void**>::kTypeId);
 					c.setArg(1, in_args);
 
-					reg params[] = { c.newGp(asmjit::Type::IdOfT<Args>::kTypeId)... };
+					// generate list of registers to bind each argument to
+					asmjit::BaseReg params[] = { reg_for_type<Args>(c)... };
+
+					// void** arg_ptr = args;
 					reg temp = c.newGp(asmjit::Type::IdOfT<void**>::kTypeId);
 					c.mov(temp, in_args);
+
+					// params[i] <- arg_ptr[i]
 					for (u8 i = 0;i < arg_count::value;i++) {
-						c.mov(params[i], asmjit::x86::Mem(temp, i * sizeof(void*)));
+						if (!arg_is_ptr[i]) {
+							if (arg_types[i] == std::type_index(typeid(f32))) {
+								c.movss(static_cast<asmjit::x86::Xmm&>(params[i]), asmjit::x86::Mem(temp, i * sizeof(void*)));
+							} else if (arg_types[i] == std::type_index(typeid(f64))) {
+								c.movsd(static_cast<asmjit::x86::Xmm&>(params[i]), asmjit::x86::Mem(temp, i * sizeof(void*)));
+							} else {
+								c.mov(static_cast<asmjit::x86::Gp&>(params[i]), asmjit::x86::Mem(temp, i * sizeof(void*)));
+							}
+						} else c.mov(static_cast<asmjit::x86::Gp&>(params[i]), asmjit::x86::Mem(temp, i * sizeof(void*)));
 					}
 
+					// original_func(...)
 					auto fcall = c.call(fimm(f), asmjit::FuncSignatureT<Ret, Args...>(asmjit::CallConv::kIdHost));
+					
+					// specify each argument
 					for (u8 i = 0;i < arg_count::value;i++) fcall->setArg(i, params[i]);
 
+					// maybe get return value
 					if constexpr (!std::is_void<Ret>::value) {
 						if (std::is_floating_point_v<Ret>) {
+							// ret = original_func(...)
 							asmjit::x86::Xmm ret = c.newXmm();
 							fcall->setRet(0, ret);
+
 							if (sizeof(Ret) == 4) {
+								//*((f32*)out_result) = ret
 								c.movss(asmjit::x86::ptr(out), ret);
 							} else if (sizeof(Ret) == 8) {
+								//*((f64*)out_result) = ret
 								c.movsd(asmjit::x86::ptr(out), ret);
 							} else {
+								//*((long f64*)out_result) = ret
 								c.movdqu(asmjit::x86::ptr(out), ret);
 							}
 						} else {
+							// ret = original_func(...)
 							reg ret = c.newGp(asmjit::Type::IdOfT<Ret>::kTypeId);
 							fcall->setRet(0, ret);
+
+							// *out_result = ret
 							c.mov(asmjit::x86::ptr(out), ret);
 						}
 					}
+
+					// return
 					c.ret();
+
+					// generate
 					c.endFunc();
 					c.finalize();
+
+					// assign func to the wrapper function
 					rt.add(&func, &h);
 				}
 
@@ -214,64 +245,108 @@ namespace gjs {
 						if (!tpm->get<Ret>()) {
 							throw bind_exception(format("Return type '%s' of method '%s' of class '%s' has not been bound yet", base_type_name<Ret>(), name.c_str(), typeid(remove_all<Cls>::type).name()));
 						}
+
+						// describe the function for the wrapped_function interface
 						ret_is_ptr = std::is_reference_v<Ret> || std::is_pointer_v<Ret>;
 						arg_is_ptr = { true, (std::is_reference_v<Args> || std::is_pointer_v<Args>)... };
 						address = *(u64*)reinterpret_cast<void*>(&f);
 
+						// begin generating the wrapper function
 						asmjit::CodeHolder h;
 						h.init(rt.codeInfo());
 						compiler c(&h);
 
+						// define function signature of the call_class_method function
 						asmjit::FuncSignatureBuilder fs;
 						fs.addArg(asmjit::Type::kIdIntPtr); // method pointer
 						fs.addArgT<Cls*>(); // self
 
+						// the rest of the argument types
 						std::vector<u32> argTypes = { asmjit::Type::IdOfT<Args>::kTypeId... };
 						for (u8 i = 0;i < ac::value;i++) fs.addArg(argTypes[i]);
+
+						// call convention and return type
 						fs.setCallConv(asmjit::CallConv::kIdHost);
 						fs.setRetT<Ret>();
 
-						c.addFunc(asmjit::FuncSignatureT<void, Ret*, void**>(asmjit::CallConv::kIdHost));
+						// define function signature 'void wrapped(void* out_result, void** args)'
+						c.addFunc(asmjit::FuncSignatureT<void, void*, void**>(asmjit::CallConv::kIdHost));
+
+						// bind out_result to register
 						reg out = c.newGp(asmjit::Type::IdOfT<Ret*>::kTypeId);
-						reg in_args = c.newGp(asmjit::Type::IdOfT<void**>::kTypeId);
 						c.setArg(0, out);
+
+						// bind args to register
+						reg in_args = c.newGp(asmjit::Type::IdOfT<void**>::kTypeId);
 						c.setArg(1, in_args);
 
-						reg params[] = {
+						// generate list of registers to bind each argument to
+						asmjit::BaseReg params[] = {
+							// this_obj
 							c.newGp(asmjit::Type::IdOfT<Cls*>::kTypeId),
-							c.newGp(asmjit::Type::IdOfT<Args>::kTypeId)...
+							// the other args
+							reg_for_type<Args>(c)...
 						};
+
+						// void** arg_ptr = args;
 						reg temp = c.newGp(asmjit::Type::IdOfT<void**>::kTypeId);
 						c.mov(temp, in_args);
+
+						// params[i] <- arg_ptr[i]
 						for (u8 i = 0;i < ac::value + 1;i++) {
-							c.mov(params[i], asmjit::x86::Mem(temp, i * sizeof(void*)));
+							if (!arg_is_ptr[i]) {
+								if (arg_types[i] == std::type_index(typeid(f32))) {
+									c.movss(static_cast<asmjit::x86::Xmm&>(params[i]), asmjit::x86::Mem(temp, i * sizeof(void*)));
+								} else if (arg_types[i] == std::type_index(typeid(f64))) {
+									c.movsd(static_cast<asmjit::x86::Xmm&>(params[i]), asmjit::x86::Mem(temp, i * sizeof(void*)));
+								} else {
+									c.mov(static_cast<asmjit::x86::Gp&>(params[i]), asmjit::x86::Mem(temp, i * sizeof(void*)));
+								}
+							} else c.mov(static_cast<asmjit::x86::Gp&>(params[i]), asmjit::x86::Mem(temp, i * sizeof(void*)));
 						}
 
+						// call_class_method<Ret, Cls, Args...>(...)
 						auto fcall = c.call(asmjit::imm(original_func), fs);
-						fcall->setArg(0, asmjit::imm(*reinterpret_cast<intptr_t*>(&f)));
+
+						// specify each argument
+						fcall->setArg(0, asmjit::imm(*reinterpret_cast<intptr_t*>(&f))); // method ptr
 						for (u8 i = 0;i <= ac::value;i++) fcall->setArg(i + 1, params[i]);
 
+						// maybe get return value
 						if constexpr (!std::is_void<Ret>::value) {
 							if (std::is_floating_point_v<Ret>) {
+								// ret = original_func(...)
 								asmjit::x86::Xmm ret = c.newXmm();
 								fcall->setRet(0, ret);
+
 								if (sizeof(Ret) == 4) {
+									//*((f32*)out_result) = ret
 									c.movss(asmjit::x86::ptr(out), ret);
 								} else if (sizeof(Ret) == 8) {
+									//*((f64*)out_result) = ret
 									c.movsd(asmjit::x86::ptr(out), ret);
 								} else {
+									//*((long f64*)out_result) = ret
 									c.movdqu(asmjit::x86::ptr(out), ret);
 								}
 							} else {
+								// ret = original_func(...)
 								reg ret = c.newGp(asmjit::Type::IdOfT<Ret>::kTypeId);
 								fcall->setRet(0, ret);
+
+								// *out_result = ret
 								c.mov(asmjit::x86::ptr(out), ret);
 							}
 						}
 
+						// return
 						c.ret();
+
+						// generate
 						c.endFunc();
 						c.finalize();
+
+						// assign func to the wrapper function
 						rt.add(&func, &h);
 					}
 
@@ -287,6 +362,9 @@ namespace gjs {
 			};
 		};
 
+		/*
+		 * Function wrapping helper
+		 */
 		template <typename Ret, typename... Args>
 		wrapped_function* wrap(type_manager* tpm, asmjit::JitRuntime& rt, const std::string& name, Ret(*func)(Args...)) {
 			return new x86::global_function<Ret, Args...>(tpm, rt, func, name);
@@ -297,6 +375,10 @@ namespace gjs {
 			return new x86::class_method<Ret, Cls, Args...>(tpm, rt, func, name);
 		};
 
+
+		/*
+		 * Class wrapping helper
+		 */
 		template <typename Cls, typename... Args>
 		Cls* construct_object(void* mem, Args... args) {
 			return new (mem) Cls(args...);
@@ -321,38 +403,8 @@ namespace gjs {
 			pf_none				= 0b00000000,
 			pf_read_only		= 0b00000001,
 			pf_write_only		= 0b00000010,
-			pf_object_pointer	= 0b00000100,
-			pf_static_prop		= 0b00001000
-		};
-
-		struct wrapped_class {
-			struct property {
-				property(wrapped_function* g, wrapped_function* s, std::type_index t, u64 o, u8 f) :
-					getter(g), setter(s), type(t), offset(o), flags(f) { }
-
-				wrapped_function* getter;
-				wrapped_function* setter;
-				std::type_index type;
-				u64 offset;
-				u8 flags;
-			};
-
-			wrapped_class(asmjit::JitRuntime& _rt, const std::string& _name, const std::string& _internal_name, size_t _size) :
-				rt(_rt), name(_name), internal_name(_internal_name), size(_size), ctor(nullptr), dtor(nullptr), requires_subtype(false)
-			{
-			}
-
-			~wrapped_class();
-
-			asmjit::JitRuntime& rt;
-			std::string name;
-			std::string internal_name;
-			bool requires_subtype;
-			wrapped_function* ctor;
-			robin_hood::unordered_map<std::string, wrapped_function*> methods;
-			robin_hood::unordered_map<std::string, property*> properties;
-			wrapped_function* dtor;
-			size_t size;
+			pf_pointer			= 0b00000100,
+			pf_static			= 0b00001000
 		};
 
 		template <typename Cls>
@@ -414,7 +466,7 @@ namespace gjs {
 					throw bind_exception(format("Attempting to bind property of type '%s' that has not been bound itself", typeid(remove_all<T>::type).name()));
 				}
 
-				properties[_name] = new property(nullptr, nullptr, typeid(remove_all<T>::type), (u64)member, flags | pf_static_prop);
+				properties[_name] = new property(nullptr, nullptr, typeid(remove_all<T>::type), (u64)member, flags | pf_static);
 				return *this;
 			}
 
@@ -445,6 +497,14 @@ namespace gjs {
 			type_manager* types;
 		};
 
+
+
+		template <typename T>
+		void to_reg(vm_context* context, const T& in, u64* reg_ptr);
+
+		template <typename T>
+		void from_reg(vm_context* context, T* out, u64* reg_ptr);
+
 		declare_input_binding(integer, context, in, reg_ptr);
 		declare_output_binding(integer, context, out, reg_ptr);
 		declare_input_binding(decimal, context, in, reg_ptr);
@@ -453,98 +513,7 @@ namespace gjs {
 		template <typename Arg, typename... Args>
 		void set_arguments(vm_context* ctx, vm_function* func, u8 idx, const Arg& a, Args... rest);
 		void set_arguments(vm_context* ctx, vm_function* func, u8 idx);
-	};
 
-	class vm_function {
-		public:
-			vm_function(vm_context* ctx, const std::string name, address addr);
-			vm_function(type_manager* mgr, vm_type* tp, bind::wrapped_function* wrapped, bool is_ctor = false);
-
-			void arg(vm_type* type);
-
-			std::string name;
-			bool is_host;
-
-			struct {
-				std::string text;
-				vm_type* return_type;
-				bool returns_on_stack;
-				vm_register return_loc;
-				std::vector<vm_type*> arg_types;
-				std::vector<vm_register> arg_locs;
-				bool is_thiscall;
-				bool returns_pointer;
-				bool is_subtype_obj_ctor;
-			} signature;
-
-			union {
-				bind::wrapped_function* wrapped;
-				address entry;
-			} access;
-
-			template <typename Ret, typename... Args>
-			void call(Ret* result, Args... args) {
-				if (sizeof...(args) != signature.arg_locs.size()) {
-					throw runtime_exception(format(
-						"Function '%s' takes %d arguments, %d %s provided",
-						name.c_str(),
-						signature.arg_locs.size(),
-						sizeof...(args),
-						(sizeof...(args)) == 1 ? "was" : "were"
-					));
-				}
-
-				if (signature.arg_locs.size() > 0) bind::set_arguments(m_ctx, this, 0, args...);
-				if (is_host) m_ctx->vm()->call_external(access.wrapped->address);
-				else m_ctx->vm()->execute(*m_ctx->code(), access.entry);
-
-				if (signature.return_type->size != 0) {
-					bind::from_reg(m_ctx, result, &m_ctx->state()->registers[(u8)signature.return_loc]);
-				}
-			}
-
-		protected:
-			vm_context* m_ctx;
-	};
-
-	class vm_type {
-		public:
-			std::string name;
-			std::string internal_name;
-			u32 size;
-			bool is_primitive;
-			bool is_unsigned;
-			bool is_floating_point;
-			bool is_builtin;
-			bool requires_subtype;
-
-			struct property {
-				u8 flags;
-				std::string name;
-				vm_type* type;
-				u64 offset;
-				vm_function* getter;
-				vm_function* setter;
-			};
-
-			vm_type* base_type;
-			vm_type* sub_type;
-			std::vector<property> properties;
-			vm_function* constructor;
-			vm_function* destructor;
-			std::vector<vm_function*> methods;
-
-			inline u32 id() const { return m_id; }
-
-		protected:
-			friend class type_manager;
-			bind::wrapped_class* m_wrapped;
-			u32 m_id;
-			vm_type();
-			~vm_type();
-	};
-
-	namespace bind {
 		template <typename T>
 		void set_argument(vm_context* ctx, vm_function* func, u8 idx, const T& arg) {
 			vm_state* state = ctx->state();
